+++ src/loaders/types.js
@@ -3,8 +3,8 @@ import JsonLoader from './JsonLoader';
 import MatrixZarrLoader from './MatrixZarrLoader';
 import GenesJsonAsMatrixZarrLoader from './GenesJsonAsMatrixZarrLoader';
 import ClustersJsonAsMatrixZarrLoader from './ClustersJsonAsMatrixZarrLoader';
-import RasterLoader from './RasterLoader';
-
+import RasterJsonLoader from './RasterJsonLoader';
+import CellSetsJsonLoader from './CellSetsJsonLoader';
 
 export const fileTypeToLoader = {
   'expression-matrix.zarr': MatrixZarrLoader,
@@ -14,6 +14,6 @@ export const fileTypeToLoader = {
   'factors.json': JsonLoader,
   'molecules.json': JsonLoader,
   'neighborhoods.json': JsonLoader,
-  'raster.json': RasterLoader,
-  'cell-sets.json': JsonLoader,
+  'raster.json': RasterJsonLoader,
+  'cell-sets.json': CellSetsJsonLoader,
 };