+++ CHANGELOG.md
@@ -39,7 +39,7 @@
     - Eliminated usage of randomly-generated uuids for layer and channel states.
 - Changed the `Spatial` component implementation (back) to a React class component, after experiencing difficult-to-debug performance issues with the function component implementation.
 - Simplified the `GenesSubscriber` implementation such that it no longer handles converting gene expression data to colors. Instead, the plot components (spatial, scatterplot, and heatmap) must look up the gene expression data and handle color conversion themselves, de-coupling them from `GenesSubscriber`.
-
+- Started to add error handling for loader instance `.load()` calls within component `useEffect` set up functions.